requirements.txt
diskcache~=5.6.1
ftfy~=6.1.1
musicbrainzngs~=0.7.1
mutagen~=1.46.0
pyacoustid~=1.2.2
spotipy~=2.23.0
pyacrcloud~=1.0.1